requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.7.4
google-generativeai==0.8.3
PyMuPDF==1.24.9
python-docx==1.1.2
chardet==5.2.0
slowapi==0.1.9
python-dotenv==1.0.1